+++ lib/processor/util.go
@@ -169,6 +169,7 @@ func IteratePartsWithSpan(
 		}
 		if err := iter(i, span, part); err != nil {
 			FlagErr(part, err)
+			span.SetTag("error", true)
 			span.LogFields(
 				olog.String("event", "error"),
 				olog.String("type", err.Error()),