+++ CHANGELOG.md
@@ -7,7 +7,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- More options for the `parse_log` processor.
+- Format `syslog_rfc3164` added to the `parse_log` processor.
+- New `multilevel` cache.
 
 ## 3.10.0 - 2020-02-05